Gather all academy services in right structure and add initial view tests
