+++ portal/academy/models.py
@@ -1,5 +1,3 @@
-import random
-import string
 from datetime import datetime
 
 from django.db import models
